put PyWidget & related stuff into its own file, they don't really belong to base.ts
